# storage/api-docs/api-docs-v1.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Toros y Vacas API",
        "description": "Primer Campeonato Mundial de Toros y Vacas",
        "version": "1.0"
    },
    "paths": {
        "/api/v1/game/create": {
            "post": {
                "tags": [
                    "Juego"
                ],
                "summary": "Crea un nuevo juego",
                "operationId": "e6583fb2848fb88496db2ca17b4d21ee",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "username": {
                                                "type": "string"
                                            },
                                            "age": {
                                                "type": "integer"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "username": "pepe23",
                                    "age": "23"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Juego creado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "El juego ha sido creado"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": "5"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object",
                                    "example": {
                                        "success": true,
                                        "message": "El juego ha sido creado",
                                        "data": {
                                            "id": 5
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error al validar los datos recibidos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "false"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Los datos recibidos no son válidos"
                                        },
                                        "data": {
                                            "properties": {
                                                "username": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "La longitud máxima para el nombre de usuario es de 50 carácteres"
                                                    }
                                                },
                                                "age": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "La edad debe ser un número entero"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object",
                                    "example": {
                                        "success": false,
                                        "message": "Los datos recibidos no son válidos",
                                        "data": {
                                            "username": [
                                                "El nombre de usuario solo puede contener letras y números"
                                            ],
                                            "age": [
                                                "La edad debe ser un número entero"
                                            ]
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/game/delete": {
            "delete": {
                "tags": [
                    "Juego"
                ],
                "summary": "Elimina los datos de un juego",
                "operationId": "2e98b288612c5bc1f25a533a777fd8a7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Identificador del juego",
                        "example": "6"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Juego eliminado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "El juego ha sido eliminado"
                                        }
                                    },
                                    "type": "object",
                                    "example": {
                                        "success": true,
                                        "message": "El juego ha sido eliminado"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "El juego no existe",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "false"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "No existe un juego con el identificador especificado"
                                        }
                                    },
                                    "type": "object",
                                    "example": {
                                        "success": false,
                                        "message": "No existe un juego con el identificador especificado"
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error al validar los datos recibidos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "false"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Los datos recibidos no son válidos"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "integer",
                                                        "example": "El identificador del juego debe ser un número entero"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object",
                                    "example": {
                                        "success": false,
                                        "message": "Los datos recibidos no son válidos",
                                        "data": {
                                            "id": [
                                                "El identificador del juego debe ser un número entero"
                                            ]
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}